Fix aborting manual tasks not allowing to restart them.
The check for whether a machine can start running or not had a special case to allow machines with the "manual" id to always start; however, that didn't work because the special case was checking the wrong property as a result of a typo in the code. This is now fixed to check the correct property.