Round small numbers to the nearest millionth
Rounding errors that accumulate during calculations which result in numbers less than 1 result in long strings with over 16 digits, most of which are zeroes or nines.

This is fixed by rounding such numbers to have at most 6 digits after the decimal. Rounding is used instead of truncating as truncation will make numbers like 0.500000000000001 into 0.5 but keep most of the digits of 0.49999999999999 as is, when both values should be interpreted as 0.5.