Allow changing assigned machine counts while running
Currently, the number of machines assigned to a given item can't be changed without manually stopping the machines, adjusting the count, then starting the machines. The result is 2 additional clicks every time the number of machines to use needs to be adjusted.

This is fixed by automatically stopping assigned machines before changing the assigned count, then starting them again if possible. This only applies if the machines aren't manually stopped. The net effect is that changing the number of machines assigned restarts whatever recipe they were assigned to.